requirements.txt
python-telegram-bot
PyYAML
requests
markdown
